Fix plugin schema
