+++ frontend/src/utils/validate/validatePluginIndexData.ts
@@ -4,58 +4,18 @@ import { pickBy } from 'lodash';
 import { z } from 'zod';
 
 import { DEFAULT_PLUGIN_DATA } from '@/constants/plugin';
-import {
-  HubDimension,
-  PluginAuthor,
-  PluginIndexData,
-  PluginType,
-  PluginWriterSaveLayer,
-} from '@/types';
+import { HubDimension, PluginAuthor, PluginIndexData } from '@/types';
 
 import {
   fallback,
-  getEnumSchemaArray,
   getOptionalStringArraySanitizer,
   getStringArraySanitizer,
   SanitizerMap,
-  stringArraySchema,
-  stringSchema,
 } from './helpers';
+import { pluginBaseSchema } from './schemas';
 
-export const pluginAuthorSchema = z.object({
-  name: z.string().min(1).or(fallback('')),
-  email: z.string().email().optional().or(fallback('')),
-  orcid: z.string().min(1).optional().or(fallback('')),
-});
-
-const pluginCategorySchema = z.object({
-  'Workflow step': stringArraySchema,
-  'Supported data': stringArraySchema,
-  'Image modality': stringArraySchema,
-});
-
-export const pluginIndexDataSchema = z.object({
-  authors: pluginAuthorSchema.array().or(fallback([])),
-  category: pluginCategorySchema.optional(),
-  description_content_type: stringSchema,
-  description_text: stringSchema,
-  description: stringSchema,
-  development_status: stringArraySchema,
-  display_name: stringSchema.optional(),
-  first_released: stringSchema,
-  license: stringSchema,
-  name: stringSchema,
-  // Make schema optional until available in /plugins/:name
-  total_installs: z.number().optional().default(0),
-  operating_system: stringArraySchema,
-  plugin_types: getEnumSchemaArray(PluginType),
-  python_version: stringSchema,
-  reader_file_extensions: stringArraySchema,
-  release_date: stringSchema,
-  summary: stringSchema,
-  version: stringSchema,
-  writer_file_extensions: stringArraySchema,
-  writer_save_layers: getEnumSchemaArray(PluginWriterSaveLayer),
+export const pluginIndexDataSchema = pluginBaseSchema.extend({
+  total_installs: z.number(),
 });
 
 export const PLUGIN_INDEX_DATA_SANITIZERS: SanitizerMap<PluginIndexData> = {
